feat: 以 App Router RSC 遷移首頁 index（維持 static/ISR）
- app/(site)/page.tsx = async Server Component：server 端抓取特色區塊 + 最新食譜作為初始資料
- 維持 static + 1h ISR（○）：export const revalidate=3600 + fetchHomeFeatures/fetchHomeRecipes 的 fetch 加 next.revalidate=3600
- PageSEO + organization/website structuredData → generateMetadata + JsonLd
- tabs 切換 / load-more 累加等 client 互動抽成 home-client（'use client'；router.push → next/navigation）
- 驗證：next build 綠燈、/ 為 ○ static（Revalidate 1h）
